# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: AK Golf Academy API
  description: Backend API for AK Golf Academy - Player development and coach management
  version: 2.0.0
  contact:
    name: AK Golf Academy
    email: dev@akgolf.no
servers:
  - url: /api/v1
    description: API Version 1
security:
  - bearerAuth: []
...
paths:
...
  /exercises:
    get:
      tags: [exercises]
      summary: List exercises
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of exercises
    post:
      tags: [exercises]
      summary: Create exercise
      responses:
        '201':
          description: Exercise created
  /exercises/{id}:
    get:
      tags: [exercises]
      summary: Get exercise by ID
      parameters:
        - $ref: '#/components/parameters/exerciseId'
      responses:
        '200':
          description: Exercise details
    patch:
      tags: [exercises]
      summary: Update exercise
      parameters:
        - $ref: '#/components/parameters/exerciseId'
      responses:
        '200':
          description: Exercise updated
    delete:
      tags: [exercises]
      summary: Delete exercise
      parameters:
        - $ref: '#/components/parameters/exerciseId'
      responses:
        '200':
          description: Exercise deleted
  /dashboard:
    get:
      tags: [dashboard]
      summary: Get player dashboard
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardResponse'
  /dashboard/{playerId}:
    get:
      tags: [dashboard]
      summary: Get dashboard for specific player (coach view)
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Player dashboard
  /dashboard/weekly-stats:
    get:
      tags: [dashboard]
      summary: Get weekly training stats
      parameters:
        - name: week
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Weekly stats
...
  /dashboard/goals:
    get:
      tags: [dashboard]
      summary: Get player goals
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, paused, cancelled]
      responses:
        '200':
          description: Player goals
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
...
    exerciseId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    DashboardResponse:
      type: object
      properties:
        player:
          type: object
        period:
          type: object
        todaySessions:
          type: array
        badges:
          type: array
        goals:
          type: array
        weeklyStats:
          type: object
        messages:
          type: array
        unreadCount:
          type: integer
...